mss/ai-service/pyproject.toml
[tool.poetry]
name = "ai-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.2"
requests = "^2.32.3"
torch = "^2.4.1"
torchvision = "^0.19.1"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
black = "^24.8.0"
mypy = "^1.11.2"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
